+++ api/src/main/java/org/openmrs/module/referencemetadata/reporting/reports/admissions/NumberOfAllAdmissions.java
@@ -12,7 +12,7 @@
  * Copyright (C) OpenMRS, LLC.  All Rights Reserved.
  */
 
-package org.openmrs.module.referencemetadata.reporting.reports;
+package org.openmrs.module.referencemetadata.reporting.reports.admissions;
 
 import org.openmrs.module.reporting.dataset.definition.SqlDataSetDefinition;
 import org.openmrs.module.reporting.evaluation.parameter.Mapped;
